feat: restructure DB and config API
dramatically simplify the config. all config items are treated the same
now but have an extra column called `category` for convenience/clarity

removed config.yaml. all config will be driven by the UI. this gives us
more freedom to iterate at the cost of some clarity in the structure of
the config. i don't think that's a big downside, though